style: header 배경색 추가
