+++ src/pages/auth/components/EmailStep.tsx
@@ -1,33 +1,33 @@
 import Input from "../../../components/Input";
-import Button from "../../../components/Button";
+// import Button from "../../../components/Button";
 import { StepProps } from "../../../types/authPropsTypes";
-import { useActiveStore } from "../../../zustand/isActiveStore";
-import { useEffect } from "react";
+// import { useActiveStore } from "../../../zustand/isActiveStore";
+// import { useEffect } from "react";
 
-function EmailStep({ email, isEmail, emailMsg, setStep, emailChangeHandler = () => {} }: StepProps) {
-  const { isActive, setIsActive } = useActiveStore();
+function EmailStep({ email, isEmail, emailMsg, emailChangeHandler = () => {} }: StepProps) {
+  // const { isActive, setIsActive } = useActiveStore();
 
   const baseLabelClass = "transition-all duration-300 text-[13px]";
   const visibleLabelClass = "opacity-100 translate-y-0";
   const hiddenLabelClass = "opacity-0 -translate-1";
 
-  useEffect(() => {
-    if (isEmail) {
-      setIsActive(true);
-    }
-  }, [isEmail]);
+  // useEffect(() => {
+  //   if (isEmail) {
+  //     setIsActive(true);
+  //   }
+  // }, [isEmail]);
 
-  const nextStepHandler = () => {
-    if (isEmail) {
-      setStep("비밀번호");
-      setIsActive(false);
-    }
-  };
+  // const nextStepHandler = () => {
+  //   if (isEmail) {
+  //     setStep("비밀번호");
+  //     setIsActive(false);
+  //   }
+  // };
 
   return (
     <>
-      <div className="w-full flex-col flex h-[80vh]">
-        <div className="flex-grow">
+      <div className="w-full">
+        <div className="">
         <p className="text-lg mb-4">이메일을 입력해주세요</p>
         <p
           className={`${!isEmail ? "redoe" : "text-grayoe-300"} ${
@@ -46,11 +46,11 @@ function EmailStep({ email, isEmail, emailMsg, setStep, emailChangeHandler = () 
         </p>
         </div>
       
-      <div className="w-full mt-auto">
+      {/* <div className="w-full mt-auto">
         <Button size="large" type="button" onClick={nextStepHandler} isActive={isActive}>
           다음
         </Button>
-      </div>
+      </div> */}
       </div>
     </>
   );
