+++ src/pages/auth/components/Login.tsx
@@ -62,6 +62,10 @@ function Login() {
     navigate("/signup")
   }
 
+  const kakaoLoginHandler = () => {
+    
+  }
+
   return (
     <form onSubmit={onSubmit} className="w-full grid">
       <div className="mt-[50px] mx-auto"></div>
@@ -98,7 +102,7 @@ function Login() {
           <p>|</p>
           <button>계정찾기</button>
         </div>
-        <button className="flex w-full xl:w-[89px] xl:h-[40px] py-4 justify-center items-center rounded-md bg-[#FEE500] text-[#1C1C1C] font-b1-regular mt-[40px]">
+        <button onClick={kakaoLoginHandler} className="flex w-full xl:w-[89px] xl:h-[40px] py-4 justify-center items-center rounded-md bg-[#FEE500] text-[#1C1C1C] font-b1-regular mt-[40px]">
           <img src={kakaologo} alt="kakaologo" className="w-[18px] mr-[15px]"/>카카오 로그인</button>
       </div>
     </form>
