+++ src/pages/auth/components/Nickname.tsx
@@ -1,36 +1,17 @@
-import { useEffect } from "react";
-import Button from "../../../components/Button";
 import Input from "../../../components/Input";
 import { StepProps } from "../../../types/authPropsTypes";
-import { useActiveStore } from "../../../zustand/isActiveStore";
 
 function Nickname({
   nickname,
   isNickname,
   nicknameMsg,
   nicknameChangeHandler=()=>{},
-  checkedChangeHandler=()=>{},
-  isCheckedAccept,
-  setStep
+
 }: StepProps) {
-  const { isActive, setIsActive } = useActiveStore();
 
   const baseLabelClass = "transition-all duration-300 text-[13px]";
   const visibleLabelClass = "opacity-100 translate-y-0";
   const hiddenLabelClass = "opacity-0 -translate-1";
-
-  useEffect(()=>{
-    if(isNickname && isCheckedAccept) {
-      setIsActive(true);
-    }
-  },[isNickname, isCheckedAccept])
-
-  const nextStep = () => {
-    if (isNickname && isCheckedAccept) {
-      setStep("가입완료");
-      setIsActive(false);
-    }
-  };
 
   return (
     <>
@@ -51,16 +32,6 @@ function Nickname({
         >
           {nicknameMsg}
         </p>
-      </div>
-
-      <div className="w-full">
-        <span className="mx-auto text-sm">
-          <input type="checkbox" checked={isCheckedAccept} onChange={checkedChangeHandler} className="mr-2 mb-6" />
-          개인정보 수집 및 이용에 대한 동의(필수)
-        </span>
-        <Button size="large" onClick={nextStep} isActive={isActive}>
-          가입완료
-        </Button>
       </div>
     </>
   );
