+++ src/pages/auth/components/PasswordStep.tsx
@@ -1,8 +1,8 @@
-import { useEffect } from "react";
-import Button from "../../../components/Button";
+// import { useEffect } from "react";
+// import Button from "../../../components/Button";
 import Input from "../../../components/Input";
 import { StepProps } from "../../../types/authPropsTypes";
-import { useActiveStore } from "../../../zustand/isActiveStore";
+// import { useActiveStore } from "../../../zustand/isActiveStore";
 
 function PasswordStep({
   password,
@@ -13,26 +13,26 @@ function PasswordStep({
   isConfirmPassword,
   confirmPasswordMsg,
   confirmPasswordChangeHandler=()=>{},
-  setStep
+
 }: StepProps) {
-  const { isActive, setIsActive } = useActiveStore();
+  // const { isActive, setIsActive } = useActiveStore();
 
   const baseLabelClass = "transition-all duration-300 text-[13px]";
   const visibleLabelClass = "opacity-100 translate-y-0";
   const hiddenLabelClass = "opacity-0 -translate-1";
 
-  useEffect(()=>{
-    if(isPassword && isConfirmPassword) {
-      setIsActive(true);
-    }
-  },[isPassword, isConfirmPassword])
+  // useEffect(()=>{
+  //   if(isPassword && isConfirmPassword) {
+  //     setIsActive(true);
+  //   }
+  // },[isPassword, isConfirmPassword])
 
-  const nextStepHandler = () => {
-    if (isPassword && isConfirmPassword) {
-      setStep("닉네임");
-      setIsActive(false);
-    }
-  };
+  // const nextStepHandler = () => {
+  //   if (isPassword && isConfirmPassword) {
+  //     setStep("닉네임");
+  //     setIsActive(false);
+  //   }
+  // };
 
   return (
     <>
@@ -79,11 +79,11 @@ function PasswordStep({
       </div>
       </div>
 
-      <div className="w-full mt-auto">
+      {/* <div className="w-full mt-auto">
         <Button size="large" type="button" onClick={nextStepHandler} isActive={isActive}>
           다음
         </Button>
-      </div>
+      </div> */}
     </>
   );
 }
