+++ src/pages/auth/components/Signup.tsx
@@ -1,9 +1,12 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import EmailStep from "./EmailStep";
 import PasswordStep from "./PasswordStep";
 import Nickname from "./Nickname";
 import Complete from "./Complete";
 import ProgressBar from "./ProgressBar";
+import { useActiveStore } from "../../../zustand/isActiveStore";
+import Button from "../../../components/Button";
+import { useNavigate } from "react-router-dom";
 
 function Signup() {
   const [email, setEmail] = useState<string>("");
@@ -23,6 +26,10 @@ function Signup() {
   const [confirmPasswordMsg, setConfirmPasswordMsg] = useState<string>("");
 
   const [step, setStep] = useState("이메일");
+  const { isActive, setIsActive } = useActiveStore();
+  const navigate = useNavigate();
+
+
 
   const emailChangeHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
     const { value } = e.target;
@@ -76,10 +83,36 @@ function Signup() {
     setIsCheckedAccept(e.target.checked);
   };
 
+  useEffect(() => {
+    if (isEmail || isConfirmPassword || isNickname) {
+      setIsActive(true);
+    }
+
+  }, [isEmail, isConfirmPassword, isNickname]);
+
+  const nextStepHandler = () => {
+    if (isEmail) {
+      setStep("비밀번호");
+      setIsActive(false);
+    }
+    if(isPassword && isConfirmPassword) {
+      setStep("닉네임");
+      setIsActive(false);
+    }
+    if(isNickname && isCheckedAccept) {
+      setStep("가입완료");
+      setIsActive(false);
+    }
+  };
+
+    const goToMain = () => {
+    navigate("/");
+  }
+
   return (
-    <div className="w-full h-[calc(100vh-56px)] xl:h-[calc(100vh-80px)]">
+    <div className="w-full h-[calc(100vh-56px)] xl:h-[calc(100vh-80px)] flex-col flex">
       <ProgressBar step={step}/>
-      <form className="mt-4">
+      <form className="mt-4 flex-grow flex flex-col">
         {step === "이메일" && (
           <EmailStep
             email={email}
@@ -113,7 +146,26 @@ function Signup() {
           />
         )}
         {step === "가입완료" && <Complete/>}
-
+      <div className="w-full mt-auto mb-6">
+        {step !== "가입완료" ? step === "닉네임" ? 
+        <div className="w-full">
+        <span className="mx-auto text-sm">
+          <input type="checkbox" checked={isCheckedAccept} onChange={checkedChangeHandler} className="mr-2 mb-6" />
+          개인정보 수집 및 이용에 대한 동의(필수)
+        </span>
+        <Button size="large" type="button" onClick={nextStepHandler} isActive={isActive}>
+          가입완료
+        </Button>
+        </div>
+        :
+        <Button size="large" type="button" onClick={nextStepHandler} isActive={isActive}>
+          다음
+        </Button>
+        :   
+        <Button size="large" type="button" onClick={goToMain}>시작하기</Button>
+    }
+        
+      </div>
       </form>
     </div>
   );
